Port code over to Windows (#26)

+++ src/bim_parse.c
@@ -185,7 +185,7 @@ parse_bp_position(const char *field, size_t length, pio_status_t *status)
  * @param data A state_t struct.
  */
 static void
-new_field(void *field, unsigned long int field_length, void *data)
+new_field(void *field, size_t field_length, void *data)
 {
     struct state_t *state = (struct state_t *) data;
     pio_status_t status;
